Remove use of shadow plugin (#8312)
This is now redundant, and it was causing strange errors because of
incorrect published metadata: if the consumer was using a version
of Java < 17, then there would be one compatible variant for both
compilation and runtime, the `shadowRuntimeElements` variant, which
would have the shadow jar as an artifact, but no dependency. This
causes difficult to diagnose error messages, because of missing
classes (typically, depending on `http-server-netty` wouldn't
transitively bring the Micronaut runtime).