Support arbitrary nesting of EachProperty/ConfigurationProperties beans (#8424)
This change adds a new ConfigurationPath interface and implementation that is used to keep track of the state of configuration nesting and to allow nesting of @EachProperty declarations. For nested @EachProperty declarations a composite qualifier is formed from all the names in the path.

For example given a pattern foo..bar..baz and the property foo.1st.bar.2nd.baz the resulting qualifier will be "1st-2nd".