fix: 403 for localhost http host and cors any
A malicious/compromised website can make HTTP requests to localhost. Typically, such requests would trigger a CORS preflight check which would prevent the request; however, some requests are "simple" and do not require a preflight check.

This PR changes the CORS filter to return 403 if the host is localhost and the cors is set to any, which is the default if you set:

micronaut.server.cors.enabled=true