fix: properly handle primitive types in bean import
Primitive types in constructors of imported beans where handled
improperly, an array of primitives in synthesized as a type with the
name of the primitive (e.g. byte[] becomes Lbyte; instead of `[B;`).

This is caused by a bug in JavaModelUtils.getTypeReference where the
given type element is used instead of the actual class element.