Stop using EnvUtil logback programmatic config
In #8674 we re-enabled programmatic logback configuration.

However the Logback class we used 'EnvUtil' moved in logback 1.3.0, so micronaut fails to start
if a newer version of logback is pulled in as a dependency.

To fix this, this PR removes the requirement for EnvUtil, and just performs the ServiceLoader call
from inside Micronaut itself.

Fixes #8758