Client tests use blocking and non-blocking client
For netty, the blocking client is basically the non-blocking client with a Flux wrapper.

For javanet, this is not the case, and the blocking client makes use of the blocking methods in HttpClient.

To ensure both implementations are working, this PR adds a property to the server under test as to whether it should test with a blocking or non-blocking client.