fix validation of write-only config properties (#8826)
This PR fixes the R2DBC tests which are failing.

They fail because the validator tries to validate write-only properties by reading them via an introspection. This PR disables that and instead validates via the bean argument.

Note that a further refinement would be to not generate an introspection here because it is not needed.

In addition when the new validator module is integrated the same change will need to be made to the forked module.