remove dependency to validation
This PR removes validation as a dependency of http-server-tck and http-tck

It updates to 6.3.4

the task gralde task preRelease is now successful

We need https://github.com/micronaut-projects/micronaut-validation/issues/63 to enable validation of external class such as POJO in the test tck without depending on validation.

I have disabled one test because of that.