Logback configuration defined via JAVA_TOOL_OPTIONS is ignored
Previously, we only checked the classpath for logback configuration files, and when refreshing we ignored the `logback.configurationFile` setting.

This PR checks the filesystem if the config cannot be found on the classpath

It also adds `logback.configurationFile` as an optional property that points to the location of the config for when we refresh the configuration.

`logback.configurationFile` has precedence over the existing `logger.config` property