missingBeans for EndpointsFilter
[`EndpointsFilter`](https://github.com/micronaut-projects/micronaut-core/blob/4.0.x/management/src/main/java/io/micronaut/management/endpoint/EndpointsFilter.java#L44) was marked as @Internal  in [#9357](https://github.com/micronaut-projects/micronaut-core/pull/9357).

However, Micronaut Security needs to replace/disable that filter. The change to the new filters API forced the replacement in security to be done using the new API.

I think we should either remove the @Internal or what I think is a more elegant solution, create a marker API to convey Endpoints sensitivity is handled externally and thus the filter should not be loaded.

That it is what this PR proposes.