Make Publishers.just(Throwable) behave like a Mono Register JustThrowPublisher as a single publisher, so that it gets treated like a Mono.
JustThrowPublisher is used in PublisherInterceptedMethod.handleException, which in turn is called by micronaut-validation when a Publisher-returning method throws an exception (as in the test case).

Treating JustThrowPublisher as a Mono makes RouteExecutor wait for its completion instead of just using it as the body of a streaming response. This leads to better error handling, including hitting the ExceptionHandler.

Fixes #9418