Fix O(n²) code size in dispatchOne generation Copy constructor generation was O(n²) in the number of properties, which could easily hit method size limits for dispatchOne. This patch replaces the O(n²) implementation with O(n), at the cost of branching for each property.
Another alternative would be to load all properties into constants and then use a tableswitch to replace the one changed property. But in my opinion this is not worth the effort, since perf will be about the same for small copies and with large copies, the bottleneck will be elsewhere.

It may also be worth investigating other bytecode size reductions. Even after this patch, the class size of the `Massive` introspection is still around 64K bytes. For example, the static initializer seems to be calling `Argument.of` twice for each property.

Fixes #9651