Add "functional" benchmark
This commit introduces a new test project which is a JMH benchmark testing
an application. It makes use of the Micronaut Gradle plugins (that is to
say builds an application in a similar way as for a user application)
and applies the AOT plugin.

There are 2 JMH jars being created. The standard JMH jar is using the
regular application classpath. Another jar is created with the
optimized jar.

To run the benchmark with the AOT optimized application, run:

`./gradlew :test-suite-benchmarks:optimizedJmhJar`

Then:

`./test-suite-benchmarks/build/libs/test-suite-benchmarks-jmh-all-<version>.jar`