Use the current JDK for testing, but 17 for building
Previously we added a PR to run (and compile) under Java 21

https://github.com/micronaut-projects/micronaut-core/pull/9978

This did not work as expected, as in reality this is configuring under the current JDK,
but compilation is being done by Java 17 from the ToolChain set up by the micronaut-build
plugin.

This PR modifies the test tasks to run on the JDK that is used to start the build.

But all compilation will be done via Java 17 from the toolchain.