Implement zealous block detection for HttpClient This new check throws an exception if *any* event loop thread would be blocked by a call to the netty BlockingHttpClient. This is usually a bug, as a blocked event loop can lose data.
However there are legitimate cases where this is fine, so there's also a config option to disable this detection. But all the cases in the micronaut-core test suite where this made a test fail were unintentional and potentially buggy, so most users should leave the setting as-is.

The error message provides two code remediation options. It also mentions that there's a config property for this, but doesn't explain which, because I want to discourage an "easy fix" that will leave the bug in place.