Extend MDCReactorSpec's http client read timeout
This sets the read timeout of the main http client (which is manually
instantiated) used in MDCReactorSpec to 5 minutes, which is the same
as what is set for the dependency-injected clients used elsewhere
throughout the test.

This spec was frequently failing in CI with an HTTP read timeout error,
likely due to resource starvation on the CI job runner causing slow
responses from the embedded server that is used in the test.