Test improvements
The test is updated such that the `@Error` handler returns a more
unusual status code (418) as a means of ensuring in the assertions that
the handler was definitely invoked.

A comment is added to the
`testErrorHandlerWithFluxChunkedSignaledImmediateError` test method
that explains the currently demonstrated behavior vs what would be
preferred in order to be less surprising in behavior.