Error handling for streaming reactive controllers
Error handling for streaming reactive controller methods, such as those
that use a `Flux` as the return type, is improved such that immediate
error signals in the stream that occur before any data has been written
will be routed appropriately to any user-supplied `@Error` handler
methods.

A new test specification is added to cover this scenario, and the TCK's
`ErrorHandlerFluxTest` is improved to cover the enhanced error handling.