Broken allowed-origins-regex configuration. (#9427)
* WIP: test to expose allowed-origins-regex configuration bug. https://github.com/micronaut-projects/micronaut-core/issues/9423

fixes #9423

As part of the driveby protection, we check that CORS is not configured to allow ANY Origin.

When setting the allowedOriginsRegex, the allowedOrigins stays as ANY, so the request was forbidden.

This fix checks to see if allowedOriginsRegex was configured before checking allowedOrigins as part of
the driveby protection.