Move URI validation step back to RoutingInBoundHandler (#10232)
Still configurable though.

The reason for this is that with the previous logic, filters could still see an invalid `request.getPath()`. This is a potential security issue because proxy filters that are built as described [in the guide](https://docs.micronaut.io/latest/guide/index.html#proxyClient) may forward the invalid user-controlled path without checking.

This patch moves the uri check back to its original place. This ensures it's checked before filters. It also replaces the request with a fixed path ("/") so that the filters cannot see a malicious user-controlled path.