fix: consider current package when generating bean imports
The internal bean element implementation didn't implement the methods
for the determination of the visibility of the class, which resulted in
the use of reflection for the bean definitions generated for the bean
import.

By implementing the visibility methods the generated code for the
imported bean definitions don't use reflection when the fields are
visible.