Small RequestHandler refactor (#10326)
* Small RequestHandler refactor
This PR replaces the StreamedHttpRequest passed to RequestHandler by a separate ByteBody argument. This removes the need for some instanceof checks.

This is a small part of #10131, but it makes for annoying merge conflicts, so I want to pull it into 4.3.x. It should have no functional or performance difference.

* checkstyle

* fix merge