Don't call out to micronaut-core from micronaut-graal
When we have a Graal/Kotlin/Maven project we add micronaut-graal to the Kapt annotation processor paths.

When we moved empty arrays to static variables in https://github.com/micronaut-projects/micronaut-core/pull/10539, we added a requirement for micronaut-core to be also on the classpath.

This PR breaks that requirement, so just adding micronaut-graal to the processors is enough.