Revert "Don't call out to micronaut-core from micronaut-graal"
This reverts commit 36ef6b91de6caee84917b1a3e0cadf41cb597fa5.

This PR https://github.com/micronaut-projects/micronaut-core/pull/10597 was an attempt to fix Maven, Graal and Kapt, but it was in error...

The issue wasn't that micronaut-graal was calling micronaut-core.
The issue was that the Kotlin processor had both Micronaut-core 4.3.1 and 4.3.10 in the classpath.

Due to the unique way Maven deals with resolution issues (it just picks the first one it sees), this meant that the processor couldn't see the variables added and renamed in https://github.com/micronaut-projects/micronaut-core/pull/10539.

And for Kapt, there's nothing we can do to manage this (apart from add every module in micronaut-core to the annotationProcessorPaths).

We are testing fixing it by re-releasing processors with the 4.3.10 version of core.