Define default value for Cookie Max-Age
The Cookie interface is updated with a constant value to represent an
undefined Max-Age, and its JavaDocs are updated to explicitly state
that the undefined value should not be encoded.

The CookieHttpCookieAdapter implementation is updated to explicitly set
the required undefined max age value in its constructor.

The new constant is consistent with the behavior of the Netty Cookie
implementation, and is meant to enforce consistency between it and the
HttpCookie based implementation, if only by explicitly stating the
intended contract.

The CookieFactory service definition is removed from the http module
as HttpCookieFactory already gets loaded by default (if no other
service definitions are loaded) via CookieFactory.INSTANCE. This
allows other explicit services definitions such as that in http-netty
to reliably override the default implementation.

Tests are added to verify the undefined max age value in both
implementations, and to ensure that NettyServerCookieEncoder can
correctly encode a Cookie created by the default HttpCookieFactory.