Add thread-core-ratio config option and change default to 1 We already have options to set the number of threads for event loops explicitly, but none that scale with host core count.
This PR adds a thread-core-ratio config option that allows for dynamically changing event loop thread count based on core count. Netty previously did the same thing when num-threads is 0, but this PR moves that logic to our code.

The default ratio is changed from nettys 2 to 1. A single thread per core reduces jitter and should achieve the same throughput in most scenarios.