fix(context-propagation): Allow other instrumentations than context propagation
fixes #11653

The `InstrumentedExecutorService` interface is also used in micrometer.

The current instrumentation only checks if a given executor service is
instrumented, but there is no notion of which specific instrumentation
it is instrumented with.

This means that if the micrometer instrumentation is applied first, the
context propagation instrumentation is skipped.

To solve this for context propagation, we provide an implementation of
the `InstrumentedExecutorService` and traverse the targets to find that
specific implementation. It's not perfect, but should be better than
before.