Introduce ReadBuffer API This PR adds a new ReadBuffer API that aims to replace the current ByteBuffer.
The main advantage is a clear ownership system based on Closeable instead of the optional ReferenceCounted. This simplifies user code, and gets rid of potentially slow `instanceof` checks. It's also necessary for netty 5 compatibility.

As the name suggests, a `ReadBuffer` is read-only. The ByteBuffer write methods are rarely used by micronaut projects, so they only add unnecessary complexity for the non-netty implementation. The few write use cases will be covered by the ByteBodyFactory buffer methods.

In addition to the new API, I also moved `AvailableByteBody` to use `ReadBuffers`. In the future, we will only need one implementation of AvailableByteBody based on ReadBuffer, instead of the array and netty based ones that exist right now.

My plan is to use ReadBuffers to also merge some of the more complex ByteBody implementations, or make them more efficient. For example, ReactiveByteBufferByteBody could use ReadBuffers instead of nio ByteBuffers, and BaseSharedBuffer could be simplified using ReadBuffers instead of having to accommodate both netty and nio buffers.