Upgrade to Groovy 5
This commit is a major upgrade to Groovy 5. There are some noticeable
changes around default methods handling, which makes the behavior
closer to the Java behavior. Some tests had to be adjusted, and
sometimes the visitor code itself.

This is for now using a SNAPSHOT version of Spock, so do NOT merge
yet.