Introduce a Micronaut module descriptor
This commit introduces a facility to list the Micronaut modules
which are found at runtime. This will be useful diagnostics
information, which can for example be used at runtime by the
Micronaut Control Panel.

It is implemented 2-fold. First of all, an independent module
named `micronaut-module-info`, standalone without any dependencies,
defines the base interface and abstract class that can be used
for such descriptors. The goal is to load the descriptors using
service loading, independently of the framework. The reason we
do this is because this module can then be used by the build
plugins to either:
   - generate an implementation of a descriptor, which will
then be loaded by Micronaut
   - statically reason about modules, for example for linting
at build time.

The second module, `micronaut-module-info-runtime`, maps these
services into Micronaut beans. A singleton, `MicronautRuntimeModules`
can be used to fetch information about modules, which are
then resolved as a hierarchy of modules.

For now, the implementation is limited, provides minimal information
such a name for the module, its version, the Maven coordinates of
the module and some tags. Eventually, we may enrich this with more
information such as the dependencies of the module, or compatibility
with other modules. For example this could become possible to express
runtime requirements such as "module A requires at least one implementation
of B at runtime, but not more than 1".

Or, it could become possible for a module to explain which test resources
modules it requires, or even supply a list of MCP servers that could be
used for coding with AI...