Refactor Pyronaut Gradle task
This commit refactors the setup of the Pyronaut Gradle compilation
task. It is somehow a prototype of what the user facing plugin
could look like. It properly registers source directory sets for
each source set, so that there's a compilation task for each
source set.

The task makes the difference between the classpath used by the
compiler itself and the classpath available for compilation
(even though in practice, both are currently using the same
classloader).

It's worth noting that the task cannot be made cacheable yet,
because the generated sources contain absolute paths (in
particular the generated entry point), so compilation is not
relocatable (this would have to be fixed at some point).

In addition, the compiler accepts a directory as an input,
but it should probably accept a list of files instead.