Add support for less buggy multipart decoder The standard netty multipart decoder is quite buggy. A new API is available at https://github.com/netty-contrib/codec-multipart . That implementation also has compatible drop-in replacement for the old netty APIs.
This PR checks whether the new implementation is available on the classpath, and if so, uses that instead of standard netty. There is also a config option to enable "quirks" that restore various bugs of the old implementation for compatibility.

This is experimental and undocumented because the new API is in alpha. I've added it to 4.10 to let people resolve long-standing issues with form data parsing.